requirements.txt
mysql-connector-python==8.0.33
redis==4.5.4